fix video fetching
