+++ src/PreviewMoment.tsx
@@ -1,22 +1,17 @@
 import './PreviewMoment.css';
 import "./AllStyles.css"
-import { useEffect, useState } from "react";
-import { getCurrentUser } from 'aws-amplify/auth';
-import { list, getUrl } from 'aws-amplify/storage';
 import { useNavigate } from "react-router-dom"
 
 interface ModalProps {
   isOpen: boolean;
+  moment: string | undefined;
   onClose: () => void;
   onRedo: () => void;
   onSave: () => void;
 }
 
-const Modal: React.FC<ModalProps> = ({ isOpen, onClose, onRedo, onSave }) => {
+const Modal: React.FC<ModalProps> = ({ isOpen, moment, onClose, onRedo, onSave }) => {
 
-  const [userID, setUserID] = useState<string | null>(null);
-  const [moment, setMoment] = useState<string | undefined>(undefined);
-   
   const navigate = useNavigate();
 
   const handleRedo = () => {
@@ -30,46 +25,7 @@ const Modal: React.FC<ModalProps> = ({ isOpen, onClose, onRedo, onSave }) => {
     });
   };
 
-  useEffect(() => {
-      if (!isOpen) return;
-      fetchUser();
-      fetchLatestVideo();
-    });
-
   if (!isOpen) return null;
-  
-  const fetchUser = async () => {
-      try {
-          const user = await getCurrentUser();
-          setUserID(user.userId);
-      } catch (error) {
-          console.error("Error fetching user:", error);
-      }
-  };
-
-  const fetchLatestVideo = async () => {
-    try {
-        const { items: videoResults } = await list({ path: `user-media/${userID}/moments/` });
-
-        if (!videoResults.length) {
-            setMoment("");
-            return;
-        }
-
-        const sortedVideos = videoResults
-            .filter(file => file?.lastModified)
-            .sort((a, b) =>
-              new Date(b?.lastModified ?? 0).getTime() - new Date(a?.lastModified ?? 0).getTime()
-            );
-            
-        const latestVideo = sortedVideos[0];
-        const urlOutput = await getUrl({ path: latestVideo.path });
-        
-        setMoment(urlOutput.url.toString());
-    } catch (error) {
-        console.error("Error fetching latest video:", error);
-    }
-};
 
   return (
     <div className="modal-overlay">
